argparse: Move argument parser helper into the cmd subpackage
