+++ planex/cmd/createmockconfig.py
@@ -13,7 +13,7 @@ import StringIO
 import yum
 import argcomplete
 
-from planex.util import add_common_parser_options
+from planex.cmd.args import add_common_parser_options
 from planex.util import setup_logging
 from planex.util import setup_sigint_handler
