+++ planex/cmd/depend.py
@@ -9,7 +9,7 @@ import sys
 import urlparse
 
 import argcomplete
-from planex.util import add_common_parser_options
+from planex.cmd.args import add_common_parser_options
 from planex.util import setup_sigint_handler
 from planex.cmd import manifest
 import planex.spec as pkg
