+++ planex/cmd/fetch.py
@@ -14,7 +14,7 @@ import pkg_resources
 import pycurl
 
 from planex.link import Link
-from planex.util import add_common_parser_options
+from planex.cmd.args import add_common_parser_options
 from planex.util import run
 from planex.util import setup_logging
 from planex.util import setup_sigint_handler
