+++ planex/cmd/init.py
@@ -10,7 +10,7 @@ import sys
 import argcomplete
 from pkg_resources import resource_filename
 
-from planex.util import add_common_parser_options
+from planex.cmd.args import add_common_parser_options
 from planex.util import setup_sigint_handler
