+++ planex/cmd/mock.py
@@ -12,7 +12,7 @@ from uuid import uuid4
 
 import argparse
 import argcomplete
-from planex.util import add_common_parser_options
+from planex.cmd.args import add_common_parser_options
 
 
 def parse_args_or_exit(argv=None):
